requirements.txt
pdf2image==1.16.3
pytesseract==0.3.10
Pillow>=11.0.0,<12.0.0
opencv-python>=4.8.0,<5.0.0
numpy>=1.24.0,<2.0.0